Default CSS used
Basically, if you choose a theme with a variation and then switch to a
completly different theme it tries to use the same variation with that
theme, even if it doesn't exist.

This PR always sets CSS to default when switching theme. Tested and
working fine on my server.

As discussed at http://forums.sentora.io/showthread.php?tid=89