make stats incremental, preserve days
I've added the -p flag (increment) to webalizer call so all the days of the month get showed in the stats, previously the webalizer was buggy, only the days found in the active log (before rotation) were showed. E.G. for each month only the 31st or 30th of the month were available.